test: trim low-signal coverage, flake sources, and slow suites
Cut Playwright journeys that only re-check unit-covered chrome or edge
cases, drop copy/SQL/Wrangler pins that add little behavior signal, and
shrink the slowest workers/bundler and migration suites so validate stays
faster and less flake-prone.
