+++ e2e/admin-feature-flags.spec.ts
@@ -33,11 +33,6 @@ test('admin feature flags: global toggle and per-user override visibility', asyn
 		page.getByRole('heading', { name: 'Admin feature flags' }),
 	).toBeVisible()
 
-	// The permanent demo flag is intentionally unmeasured.
-	await expect(
-		page.getByTestId('success-metric-notice-demo-indicator'),
-	).toContainText('No success metric declared')
-
 	const demoFlagSection = page
 		.getByRole('heading', { name: 'demo-indicator' })
 		.locator('xpath=ancestor::section[1]')
