+++ packages/worker/src/mcp/capabilities/build-capability-registry.node.test.ts
