+++ packages/worker/src/mcp/capabilities/capability-search.node.test.ts
@@ -1,4 +1,5 @@
 import { expect, test } from 'vitest'
+import { lexicalScore } from '#worker/vectorize/scoring.ts'
 import {
 	CAPABILITY_EMBEDDING_DIMENSIONS,
 	deterministicEmbedding,
@@ -133,4 +134,61 @@ test('capability Vectorize stays builtin-scoped, query-embed-only, and lexical-o
 	expect(lexicalOnly?.vectorScore).toBe(0)
 	expect(lexicalOnly!.lexicalScore).toBeGreaterThan(vectorHit!.lexicalScore)
 	expect(vectorHit?.vectorRank).toBe(1)
+})
+
+test('offline capability search ranks lexical matches and returns structured detail without schema fields', async () => {
+	const doc = 'alpha beta gamma delta epsilon'
+	expect(lexicalScore('alpha beta', doc)).toBeGreaterThan(
+		lexicalScore('omega zeta', doc),
+	)
+
+	const specs = {
+		oauth_setup_guide: {
+			name: 'oauth_setup_guide',
+			domain: 'coding',
+			description: 'Guide for configuring OAuth redirect URIs.',
+			keywords: ['oauth', 'redirect uri', 'provider registration'],
+			readOnly: true,
+			idempotent: true,
+			destructive: false,
+			source: 'builtin',
+			inputFields: ['guide'],
+			requiredInputFields: ['guide'],
+			outputFields: ['title', 'body'],
+			inputSchema: {
+				type: 'object',
+				properties: {
+					guide: { type: 'string' },
+				},
+				required: ['guide'],
+			},
+			inputTypeDefinition: 'type OAuthSetupGuideInput = {\n\tguide: string\n}',
+		},
+	} satisfies Record<string, CapabilitySpec>
+	const env = {
+		SENTRY_ENVIRONMENT: 'test',
+		AI: {} as Ai,
+	} as Env
+
+	const oauthGuide = await searchCapabilities({
+		env,
+		query: 'oauth redirect uri provider registration',
+		limit: 8,
+		detail: true,
+		specs,
+	})
+
+	expect(oauthGuide.offline).toBe(true)
+	expect(oauthGuide.matches).toHaveLength(1)
+	expect(oauthGuide.matches[0]).toMatchObject({
+		name: 'oauth_setup_guide',
+		outputFields: ['title', 'body'],
+		lexicalRank: 1,
+		vectorRank: 1,
+	})
+	expect(oauthGuide.matches[0]).not.toHaveProperty('inputSchema')
+	expect(oauthGuide.matches[0]).toMatchObject({
+		inputTypeDefinition: expect.stringContaining('OAuthSetupGuideInput'),
+	})
+	expect(oauthGuide.matches[0]).not.toHaveProperty('outputSchema')
 })
